backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
groq
pypdf
python-docx
python-pptx
pandas
python-dotenv
websockets
